cpp08 ex01: fix signed overflow in Span::shortestSpan/longestSpan
int - int can't represent the true span when the numbers include
values near both ends of int's range - e.g. INT_MAX - INT_MIN is
4294967295, which overflows a 32-bit signed int (undefined behavior,
silently wrapped to -1 on this compiler instead of crashing).

long doesn't fix it either on this platform: sizeof(long) == sizeof(int)
== 4 here (same lesson learned earlier with ScalarConverter's overflow
bug). Switched both functions to long long (8 bytes, confirmed via
sizeof), computing every difference with explicit static_cast<long long>
before subtracting so the overflow never happens even as an intermediate
step. Added a permanent INT_MAX/INT_MIN test to main.cpp proving the
correct value (4294967295) instead of the previous silent -1.

+++ cpp08/ex01/Span.cpp
@@ -35,7 +35,7 @@ void Span::addNumber(int value)
     _numbers.push_back(value);
 }
 
-int Span::shortestSpan() const
+long long Span::shortestSpan() const
 {
     if (_numbers.size() < 2)
         throw std::length_error("Not enough numbers to find a span");
@@ -43,22 +43,22 @@ int Span::shortestSpan() const
     std::vector<int> sorted(_numbers);
     std::sort(sorted.begin(), sorted.end());
 
-    int shortest = sorted[1] - sorted[0];
+    long long shortest = static_cast<long long>(sorted[1]) - static_cast<long long>(sorted[0]);
     for (std::vector<int>::size_type i = 2; i < sorted.size(); ++i)
     {
-        int diff = sorted[i] - sorted[i - 1];
+        long long diff = static_cast<long long>(sorted[i]) - static_cast<long long>(sorted[i - 1]);
         if (diff < shortest)
             shortest = diff;
     }
     return (shortest);
 }
 
-int Span::longestSpan() const
+long long Span::longestSpan() const
 {
     if (_numbers.size() < 2)
         throw std::length_error("Not enough numbers to find a span");
 
     std::vector<int>::const_iterator minIt = std::min_element(_numbers.begin(), _numbers.end());
     std::vector<int>::const_iterator maxIt = std::max_element(_numbers.begin(), _numbers.end());
-    return (*maxIt - *minIt);
+    return (static_cast<long long>(*maxIt) - static_cast<long long>(*minIt));
 }
